+++ docs/CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.1.10] - 2023-09-11
+### Added
+- `Cursor` class
+  - private `getRangeValuesPoint ( )` to get the center point of the value of a range element
+  - private `getValueFromStandardDeviation ( )` to get the standard deviation range value
+
+### Changed
+- `Cursor` class, refactored `distance ( )` and `angle ( )` setters, and `nextElement ( )` public function, to take into account range input devices
+
 ## [0.1.9] - 2023-09-07
 ### Changed
 - `Cursor` class, refactored and finalized public `createSelectOptions ( )` to recreate a select input element and mirror its inputs
@@ -133,7 +142,8 @@ All notable changes to this project will be documented in this file.
 ---
 | Version | Date       | Commit                                                                   | Comments                                                          |
 | :-----: | :--------: | :----------------------------------------------------------------------: | :---------------------------------------------------------------- |
-| 0.1.9   | 2023-09-07 | CURRENT REVISION                                                         | Finalized public createSelectOptions ( ) to recreate a select element while mirroring its inputs; for the Cursor class
+| 0.1.10  | 2023-09-11 | CURRENT REVISION                                                         | Further refinements to the Cursor class, for range input elements
+| 0.1.9   | 2023-09-07 | [63eee86](https://github.com/Justin-Byrne/MouseMove/commit/63eee86)      | Finalized public createSelectOptions ( ) to recreate a select element while mirroring its inputs; for the Cursor class
 | 0.1.8   | 2023-09-06 | [826267b](https://github.com/Justin-Byrne/MouseMove/commit/826267b)      | Added Text class, with minor additions to Pattern, Cursor, and MouseMove classes
 | 0.1.7   | 2023-08-25 | [5c625d7](https://github.com/Justin-Byrne/MouseMove/commit/5c625d7)      | Minor refactoring and cleanup of Cursor and MouseMove classes
 | 0.1.6   | 2023-08-24 | [59aa038](https://github.com/Justin-Byrne/MouseMove/commit/59aa038)      | Refactored Cursor class's cursor type settings while including more cursor types